Move meta tests into infrastructure/tooling
